# apps/shared/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "документация для api для mia coffee bot",
    "version": "1.0.0",
    "description": "Описание api для демонстрации документации в формате JSON"
  },
  "paths": {
...
    "/api/user/favorite": {
      "get": {
        "description": "роут для получения всей продукции, которую пользователь добавил в избранное",
        "responses": {
          "200": {
            "description": "успешное получения всей продукции, добавленной в избранное",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "роут для добавления определённого товара в избранное",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "itemId": {
                    "type": "number"
                  }
                },
                "required": ["itemId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "успешное добавление товара в избранное",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "роут для удаления предмета из избранного",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "itemId": {
                    "type": "number"
                  }
                },
                "required": ["itemId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "успешное выполнение роута",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ShortItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "favorite": {
            "type": "boolean"
          }
        },
        "required": ["id", "title", "image", "favorite"]
      },
...
    }
  }
}